+++ moe/tests/views/rest/bandit_epsilon_test.py
@@ -4,8 +4,6 @@ import pyramid.testing
 
 import simplejson as json
 
-import testify as T
-
 from moe.bandit.constant import BANDIT_EPSILON_ENDPOINT, EPSILON_SUBTYPES_TO_DEFAULT_HYPERPARAMETER_INFOS, EPSILON_SUBTYPE_FIRST, EPSILON_SUBTYPE_GREEDY
 from moe.tests.bandit.bandit_test_case import BanditTestCase
 from moe.tests.views.rest.bandit_test import TestBanditViews
@@ -22,7 +20,8 @@ class TestBanditEpsilonViews(TestBanditViews):
     _moe_route = BANDIT_EPSILON_MOE_ROUTE
     _view = BanditEpsilonView
 
-    def _build_json_payload(self, subtype, historical_info, hyperparameter_info=None):
+    @staticmethod
+    def _build_json_payload(subtype, historical_info, hyperparameter_info=None):
         """Create a json_payload to POST to the /bandit/epsilon endpoint with all needed info."""
         if hyperparameter_info is None:
             hyperparameter_info = EPSILON_SUBTYPES_TO_DEFAULT_HYPERPARAMETER_INFOS[subtype]
@@ -46,7 +45,7 @@ class TestBanditEpsilonViews(TestBanditViews):
         view = BanditEpsilonView(request)
         params = view.get_params_from_request()
 
-        T.assert_dicts_equal(params['hyperparameter_info'], json_payload['hyperparameter_info'])
+        assert params['hyperparameter_info'] == json_payload['hyperparameter_info']
 
         # Test arbitrary epsilons get passed through
         json_payload['hyperparameter_info']['epsilon'] = 1.0
@@ -56,7 +55,7 @@ class TestBanditEpsilonViews(TestBanditViews):
         view = BanditEpsilonView(request)
         params = view.get_params_from_request()
 
-        T.assert_dicts_equal(params['hyperparameter_info'], json_payload['hyperparameter_info'])
+        assert params['hyperparameter_info'] == json_payload['hyperparameter_info']
 
     def test_epsilon_first_hyperparameters_passed_through(self):
         """Test that the hyperparameters get passed through to the epsilon-first endpoint."""
@@ -70,7 +69,7 @@ class TestBanditEpsilonViews(TestBanditViews):
         view = self._view(request)
         params = view.get_params_from_request()
 
-        T.assert_dicts_equal(params['hyperparameter_info'], json_payload['hyperparameter_info'])
+        assert params['hyperparameter_info'] == json_payload['hyperparameter_info']
 
         # Test an arbitrary epsilon and total_tamples get passed through
         json_payload['hyperparameter_info']['epsilon'] = 1.0
@@ -81,7 +80,7 @@ class TestBanditEpsilonViews(TestBanditViews):
         view = self._view(request)
         params = view.get_params_from_request()
 
-        T.assert_dicts_equal(params['hyperparameter_info'], json_payload['hyperparameter_info'])
+        assert params['hyperparameter_info'] == json_payload['hyperparameter_info']
 
     def test_historical_info_passed_through(self):
         """Test that the historical info get passed through to the endpoint."""
@@ -90,7 +89,3 @@ class TestBanditEpsilonViews(TestBanditViews):
     def test_interface_returns_as_expected(self):
         """Integration test for the /bandit/epsilon endpoint."""
         self._test_interface_returns_as_expected()
-
-
-if __name__ == "__main__":
-    T.run()