+++ moe/tests/views/rest/gp_ei_test.py
@@ -4,8 +4,6 @@ import numpy
 
 import simplejson as json
 
-import testify as T
-
 from moe.optimal_learning.python.cpp_wrappers.covariance import SquareExponential
 from moe.optimal_learning.python.cpp_wrappers.expected_improvement import ExpectedImprovement
 from moe.optimal_learning.python.cpp_wrappers.gaussian_process import GaussianProcess
@@ -22,7 +20,8 @@ class TestGpEiView(GaussianProcessTestCase, RestTestCase):
     precompute_gaussian_process_data = True
     endpoint = GP_EI_ENDPOINT
 
-    def _build_json_payload(self, domain, covariance, historical_data, points_to_evaluate):
+    @staticmethod
+    def _build_json_payload(domain, covariance, historical_data, points_to_evaluate):
         """Create a json_payload to POST to the /gp/ei endpoint with all needed info."""
         json_payload = json.dumps({
             'points_to_evaluate': points_to_evaluate,
@@ -71,7 +70,3 @@ class TestGpEiView(GaussianProcessTestCase, RestTestCase):
                     cpp_expected_improvement,
                     tolerance,
                     )
-
-
-if __name__ == "__main__":
-    T.run()