+++ moe_examples/tests/moe_example_test_case.py
@@ -1,16 +1,17 @@
 # -*- coding: utf-8 -*-
 """Base class for testing the moe examples."""
-import testify as T
+import pytest
 
 
-class MoeExampleTestCase(T.TestCase):
+class MoeExampleTestCase(object):
 
     """Base class for testing the moe examples."""
 
-    @T.class_setup
-    def create_webapp(self):
-        """Create a mocked webapp and store it in self.testapp."""
+    @classmethod
+    @pytest.fixture(autouse=True, scope='class')
+    def create_webapp(cls):
+        """Create a mocked webapp and store it in cls.testapp."""
         from moe import main
         app = main({}, use_mongo='false')
         from webtest import TestApp
-        self.testapp = TestApp(app)
+        cls.testapp = TestApp(app)